Fixed bug in texture.c
git-svn-id: https://starruby.googlecode.com/svn/trunk@551 425b54fa-293c-0410-bdda-a7e424c12f17

+++ texture.c
@@ -981,8 +981,8 @@ Texture_render_texture(int argc, VALUE* argv, VALUE self)
     int_fast32_t srcJ16 = srcOY16 + j * srcDYY16;
     for (int i = 0; i < dstWidth;
          i++, dst++, srcI16 += srcDXX16, srcJ16 += srcDXY16) {
-      uint8_t srcI = srcI16 >> 16;
-      uint8_t srcJ = srcJ16 >> 16;
+      int_fast32_t srcI = srcI16 >> 16;
+      int_fast32_t srcJ = srcJ16 >> 16;
       if (srcX <= srcI && srcI < srcX2 && srcY <= srcJ && srcJ < srcY2) {
         Pixel src = srcTexture->pixels[srcI + srcJ * srcTextureWidth];
         if (saturation < 255) {
